requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
psycopg2-binary==2.9.9